Updated
- Fixed : database migrations files

+++ database/migrations/v1_2/2020_10_15_231045_adjust_nexopos_procurements_products_table_oct16.php
@@ -37,20 +37,24 @@ class AdjustNexoposProcurementsProductsTableOct16 extends Migration
      */
     public function down()
     {
-        Schema::table( 'nexopos_procurements_products', function( Blueprint $table ) {
-            if ( Schema::hasColumn( 'nexopos_procurements_products', 'expiration_date' ) ) {
-                $table->dropColumn( 'expiration_date' );
-            }
+        if ( Schema::hasTable( 'nexopos_procurements_products' ) ) {
+            Schema::table( 'nexopos_procurements_products', function( Blueprint $table ) {
+                if ( Schema::hasColumn( 'nexopos_procurements_products', 'expiration_date' ) ) {
+                    $table->dropColumn( 'expiration_date' );
+                }
+    
+                if ( Schema::hasColumn( 'nexopos_procurements_products', 'barcode' ) ) {
+                    $table->dropColumn( 'barcode' );
+                }
+            });
+        }
 
-            if ( Schema::hasColumn( 'nexopos_procurements_products', 'barcode' ) ) {
-                $table->dropColumn( 'barcode' );
-            }
-        });
-
-        Schema::table( 'nexopos_products', function( Blueprint $table ) {
-            if ( ! Schema::hasColumn( 'nexopos_products', 'expiration' ) ) {
-                $table->datetime( 'expiration' )->nullable();
-            }
-        });
+        if ( Schema::hasTable( 'nexopos_products' ) ) {
+            Schema::table( 'nexopos_products', function( Blueprint $table ) {
+                if ( ! Schema::hasColumn( 'nexopos_products', 'expiration' ) ) {
+                    $table->datetime( 'expiration' )->nullable();
+                }
+            });
+        }
     }
 }
